Implemented scheduler for periodical history parse task.

+++ pack.sh
@@ -1,4 +1,13 @@
-mkdir ./out &&
+service_name=$1
+version=$2
+
+if [ "$#" -ne "2" ]; then
+	echo "Service name and version aren't specified." 1>&2
+	exit 64
+fi
+
+rm -rf ./out &&
+	mkdir ./out &&
 	cp -R ./history-parse ./out/history-parse && cp -R ./history-parse-objects ./out/history-parse-objects &&
 
 	rm -rf ./out/history-parse/bin/ && rm -rf ./out/history-parse/obj/ &&
@@ -9,10 +18,14 @@ mkdir ./out &&
 	dotnet restore -r linux-musl-x64 -v m &&
 	dotnet publish -c Release -o ./../pub -r linux-musl-x64 -v m &&
 
-	cd ./../.. &&
+	cd ./../..
 
-	docker build -t history-parse:0.1.0 -f Dockerfile . && 
-	docker create --name history-parse --network cs-go-stats_prime history-parse:0.1.0 &&
-	docker start history-parse
+docker stop $service_name || true
+docker container rm $service_name || true
+docker image rm $service_name:$version || true
 
-	rm -rf ./out
+docker build -t $service_name:$version -f Dockerfile . && 
+	docker create --name $service_name --network cs-go-stats_prime $service_name:$version &&
+	docker start $service_name
+
+rm -rf ./out
